requirements.txt
# Core Dependencies
crewai>=0.28.0
crewai-tools>=0.4.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
kaggle>=1.5.0
google-ads>=23.0.0
facebook-business>=19.0.0
requests>=2.31.0
openpyxl>=3.1.0
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Testing Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Code Quality & Linting
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
pylint>=3.0.0
isort>=5.13.0

# Monitoring & Observability
prometheus-client>=0.19.0

# UI
gradio>=4.0.0

# Data Science
numpy>=1.24.0
scikit-learn>=1.3.0

# Additional Utilities
colorama>=0.4.6
